+++ opaca-platform/src/main/java/de/gtarc/opaca/platform/events/EventsFilter.java
@@ -1,10 +1,9 @@
-package de.gtarc.opaca.platform;
+package de.gtarc.opaca.platform.events;
 
 import de.gtarc.opaca.model.Event;
 import de.gtarc.opaca.util.EventHistory;
 import jakarta.servlet.*;
 import jakarta.servlet.http.*;
-import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.io.IOException;
@@ -15,7 +14,7 @@ import java.util.Set;
  * Filter for pre- and postprocessing requests. Can be used for generating Events for the
  * History, for uniform logging, or for outright rejecting certain requests.
  */
-@Service @NoArgsConstructor
+@Service
 public class EventsFilter implements Filter {
 
     @Autowired