+++ opaca-platform/src/main/java/de/gtarc/opaca/platform/session/SessionData.java
@@ -1,7 +1,9 @@
 package de.gtarc.opaca.platform.session;
 
 import java.util.*;
+import java.util.stream.Stream;
 
+import de.gtarc.opaca.model.AgentDescription;
 import de.gtarc.opaca.model.PostAgentContainer;
 import de.gtarc.opaca.platform.containerclient.DockerClient;
 import de.gtarc.opaca.platform.containerclient.KubernetesClient;
@@ -13,9 +15,9 @@ import lombok.Data;
 
 /**
  * Class aggregating all Session data of the Runtime Platform, to be stored to and loaded from
- * file in between sessions. All other classes (e.g. Runtime-Impl etc.) use the data in this class.
+ * a file in between sessions. All other classes (e.g., Runtime-Impl etc.) use the data in this class.
  */
-@Data @Component
+@Component
 public class SessionData {
 
     /* PlatformImpl variables */
@@ -23,20 +25,30 @@ public class SessionData {
     public Map<String, PostAgentContainer> startContainerRequests = new HashMap<>();
     public Map<String, RuntimePlatform> connectedPlatforms = new HashMap<>();
 
-    /* DockerClient variables */
-    public Map<String, DockerClient.DockerContainerInfo> dockerContainers = new HashMap<>();
+    /* Docker/Kubernetes containers state, for "reconnect" policy */
     public Set<Integer> usedPorts = new HashSet<>();
+    public Map<String, DockerClient.DockerContainerInfo> dockerContainers = new HashMap<>();
+    public Map<String, KubernetesClient.PodInfo> kubernetesPods = new HashMap<>();
 
-    /* KubernetesClient variables */
-    public Map<String, KubernetesClient.PodInfo> pods = new HashMap<>();
 
     public void reset() {
         this.runningContainers.clear();
         this.startContainerRequests.clear();
         this.connectedPlatforms.clear();
-        this.dockerContainers.clear();
         this.usedPorts.clear();
-        this.pods.clear();
+        this.dockerContainers.clear();
+        this.kubernetesPods.clear();
+    }
+
+    public Stream<AgentDescription> streamAgents(boolean includeConnected) {
+        return streamContainers(includeConnected).flatMap(c -> c.getAgents().stream());
+    }
+
+    public Stream<AgentContainer> streamContainers(boolean includeConnected) {
+        return includeConnected ? Stream.concat(
+                runningContainers.values().stream(),
+                connectedPlatforms.values().stream().flatMap(rp -> rp.getContainers().stream())
+        ) : runningContainers.values().stream();
     }
 
 }